+++ src/causalitytests/tests_entropy_based/ExactSimplexIntersectionTest.jl
@@ -85,7 +85,7 @@ te_ytox = causality(y, x, test)
     using the Perron-Frobenius operator." Physical Review E 99.4 (2019): 042212.
     [https://journals.aps.org/pre/abstract/10.1103/PhysRevE.99.042212](https://journals.aps.org/pre/abstract/10.1103/PhysRevE.99.042212)
 """
-Base.@kwdef struct ExactSimplexIntersectionTest{N} <: TransferEntropyCausalityTest{N}
+Base.@kwdef mutable struct ExactSimplexIntersectionTest{N} <: TransferEntropyCausalityTest{N}
     """ The delay reconstruction parameter k (controls dimension of ``T_{f}`` component of embedding). """
     k::Int = 1
 
@@ -127,7 +127,7 @@ Base.@kwdef struct ExactSimplexIntersectionTest{N} <: TransferEntropyCausalityTe
 
 
     """ The prediction lags"""
-    ηs
+    ηs::Union{Int, AbstractVector{Int}}
 
     function ExactSimplexIntersectionTest(k::Int, l::Int, m::Int, n::Int, τ::Int, b::Number, 
             estimator::E, n_pts::Int,