+++ docs/src/CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## Release v0.7.0
 
+### Breaking changes
+
+- The `ν` parameter for the `ConvergentCrossMappingTest` and `CrossMappingTest` 
+    has been changed to `η` to conform with the syntax for transfer entropy tests.
+
 ### New features
 
 - All subtypes of `CausalityTest` now have a type parameter `N` indicating
     the number of elements that are returned when applying the tests.
-- The `ν` parameter for the `ConvergentCrossMappingTest` and `CrossMappingTest` 
-    has been changed to `η`.
 
 ## Release v0.6.1